Add Make file

+++ roboshop/common.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 
+USER_ID=$(id -u)
+if [ "$USER_ID" -ne 0 ]; then
+  echo -e "\e[31mYou should be root user/sudo user to run this script\e[0m"
+  exit 2
+fi
+
 LOG=/tmp/roboshop.log
 rm -f $LOG
 
 STAT_CHECK() {
-  if [ $1 -eq 0 ]; then
+  if [ "$1" -eq 0 ]; then
     echo -e "\e[32m done\e[0m"
   else
     echo -e "\e[31m fail\e[0m"
